# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Challenge",
    "version": "v1"
  },
  "paths": {
    "/api/InfoRefrigeradores": {
      "get": {
        "tags": [
          "InfoRefrigeradores"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefrigeradoresEntity"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefrigeradoresEntity"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefrigeradoresEntity"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "InfoRefrigeradores"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefrigeradoresEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefrigeradoresEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefrigeradoresEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefrigeradoresEntity"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefrigeradoresEntity"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefrigeradoresEntity"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "InfoRefrigeradores"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefrigeradoresEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefrigeradoresEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefrigeradoresEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefrigeradoresEntity"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefrigeradoresEntity"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefrigeradoresEntity"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/InfoRefrigeradores/{id}": {
      "delete": {
        "tags": [
          "InfoRefrigeradores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefrigeradoresEntity"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefrigeradoresEntity"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefrigeradoresEntity"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RefrigeradoresEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "temperatura": {
            "type": "string",
            "nullable": true
          },
          "espacioInterno": {
            "type": "string",
            "nullable": true
          },
          "consumoInterno": {
            "type": "string",
            "nullable": true
          },
          "estadoLuces": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}